+++ can't solve Q3.playground/Contents.swift
@@ -17,6 +17,7 @@ for i in 3...100{
 }
 
 print(list)
+// 함수가 없어서 뭘 할수가 없어요.
