+++ solved Q5.playground/playground.xcworkspace/xcshareddata/IDEWorkspaceChecks.plist
