+++ unsolved Q4.playground/Contents.swift
@@ -24,5 +24,6 @@ for i in 2...100{
 
 print(list)
 
+//문제가 뭔지를 모르겠어요
